Update display.php
This display.php consists of code to display random set of songs for user to like/dislike. Every time the user visits the page he/she  will be given random set of songs for selection . We have set a minimum limit of liking as 7 out of 10 so that we can make our recommendation more better.